requirements.txt
openai
tiktoken
simplejson
nltk
python-docx